+++ CHANGELOG.md
@@ -4,6 +4,18 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a ch
 
 ## [Unreleased](https://github.com/idealista/jenkins-role/tree/develop)
 
+## [2.0.0](https://github.com/idealista/jenkins-role/tree/2.0.0)
+## [Full Changelog](https://github.com/idealista/jenkins-role/compare/1.2.0...2.0.0)
+### Changed
+- *Install default plugins* @jmonterrubio
+- *Install default settings* @jmonterrubio
+- *Remove CLI* @jmonterrubio
+
+### Added
+- *Improve security* @jmonterrubio
+- *Job DSL execution* @jmonterrubio
+
+
 ## [1.2.0](https://github.com/idealista/jenkins-role/tree/1.2.0)
 ## [Full Changelog](https://github.com/idealista/jenkins-role/compare/1.1.0...1.2.0)
 ### Changed